% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup-latex.R
\name{cleanup_latex_files}
\alias{cleanup_latex_files}
\title{Clean up LaTeX files created by exportLatex}
\usage{
cleanup_latex_files(dir = ".", outname = "", verbose = FALSE)
}
\arguments{
\item{dir}{Directory where LaTeX files are located (default: current directory)}

\item{outname}{Prefix for the LaTeX files (default: empty string)}

\item{verbose}{If TRUE, print messages about which files were removed}
}
\value{
Invisibly returns a logical vector indicating which files were successfully removed
}
\description{
This function removes LaTeX files that are typically created by the exportLatex() function.
It's designed to be used after exportLatex() calls to clean up temporary files.
}
\examples{
\dontrun{
# After calling exportLatex()
exportLatex(vim)
cleanup_latex_files()

# With custom directory and prefix
exportLatex(vim, outname = "myresults_", dir = "output/")
cleanup_latex_files(dir = "output/", outname = "myresults_")
}

}
