% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_tmle.R
\name{estimate_tmle}
\alias{estimate_tmle}
\title{Get TMLE estimate.}
\usage{
estimate_tmle(Y, A, W, family, delta = NULL, Q.lib, g.lib, verbose = F)
}
\arguments{
\item{Y}{Outcome variable}

\item{A}{Treatment indicator}

\item{W}{Dataframe of adjustment covariates}

\item{family}{Binomial or gaussian}

\item{delta}{Indicator of missing outcome or treatment assignment. 1 - observed, 0 - missing.}

\item{Q.lib}{SuperLearner library for estimating Q (potential outcome)}

\item{g.lib}{SuperLearner library for estimating g (propensity score)}

\item{verbose}{If true output extra information during execution.}
}
\description{
Get TMLE estimate.
}

