% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImpact.R
\name{exportLatex}
\alias{exportLatex}
\title{Export varImpact results as Latex tables}
\usage{
exportLatex(impact_results, outname = "", dir = ".")
}
\arguments{
\item{impact_results}{Result object from previous varImpact() call.}

\item{outname}{(Optional) String that is prepended to filenames.}

\item{dir}{(Optional) Directory to save the results, defaults to current directory.}
}
\description{
Outputs results from varImpact() into three Latex tables: consistent results,
all results, and per-fold results.
}
\details{
[Add details]
}

