% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors_to_indicators.R
\name{factors_to_indicators}
\alias{factors_to_indicators}
\title{Convert a dataframe of factor into separate indicators.}
\usage{
factors_to_indicators(factor_df, miss_name_prefix = "Imiss_",
  verbose = F)
}
\arguments{
\item{factor_df}{Dataframe consisting only of factor variables.}

\item{miss_name_prefix}{Starting name for each missing indicator.}

\item{verbose}{Set to T for detailed output.}
}
\value{
A list of results.
}
\description{
More details
}
