% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_sqr.R
\name{max_sqr}
\alias{max_sqr}
\title{Find the maximum of the squared values.}
\usage{
max_sqr(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Maximum of the squared values, or -Inf if all elements are NA.
}
\description{
Find the maximum of the squared values.
}

