% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-var.R
\name{plot_var}
\alias{plot_var}
\title{Plot the varimpact results for a given variable}
\usage{
plot_var(var_name, vim, digits = 2L, verbose = FALSE)
}
\arguments{
\item{var_name}{String name of the variable}

\item{vim}{Varimpact result object that contains the variable}

\item{digits}{Number of digits for rounding purposes.}

\item{verbose}{If true, display extra output.}
}
\description{
Displays the adjusted treatment-specific means and the impact estimate.
}
