% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImpact.R
\name{print.varImpact}
\alias{print.varImpact}
\title{Custom printing of the varImpact results.}
\usage{
\method{print}{varImpact}(x, ...)
}
\arguments{
\item{x}{Results object from varImpact.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Shows the consistent results by default. If there are no consistent results
it shows all results.
}

