% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles_equivalent.R
\name{quantiles_equivalent}
\alias{quantiles_equivalent}
\title{Checks if two quantiles have the same sample value for a given vector.
This indicates that the vector has little variation.}
\usage{
quantiles_equivalent(x, quantile_probs = c(0.1, 0.9))
}
\arguments{
\item{x}{Data vector. If a factor it is converted to numeric using unclass().}

\item{quantile_probs}{Vector with two probabilities that specify the quantiles.}
}
\value{
True if the two quantiles are equal, indicating a lack of variation in the sample data.
}
\description{
Checks if two quantiles have the same sample value for a given vector.
This indicates that the vector has little variation.
}
\seealso{
restrict_by_quantiles
}

