% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_dimensions.R
\name{reduce_dimensions}
\alias{reduce_dimensions}
\title{Reduce variables in a dataframe to a target number of covariates.}
\usage{
reduce_dimensions(data, newX = NULL, max_variables, verbose = F)
}
\arguments{
\item{data}{Dataframe}

\item{newX}{Optional second dataframe to receive the same reduction.}

\item{max_variables}{Maximum we want to allow, after which dimension reduction
will take place. Cannot be more than 15 due to HOPACH limitations. Set to NULL
to disable any dimension reduction.}

\item{verbose}{If true will output additional information during execution.}
}
\description{
Currently uses HOPACH hierarchical clustering but could be generalized.
}
