% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrict_by_quantiles.R
\name{restrict_by_quantiles}
\alias{restrict_by_quantiles}
\title{Remove columns from a dataframe if they do not have sufficient variation.}
\usage{
restrict_by_quantiles(data, quantile_probs = c(0.1, 0.9))
}
\arguments{
\item{data}{Dataframe or matrix}

\item{quantile_probs}{The probabilities corresponding to the quantiles that will be compared.}
}
\value{
New dataframe with the restriction applied.
}
\description{
Remove columns from a dataframe if they do not have sufficient variation.
}
\seealso{
quantiles_equivalent
}

