% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_factors_numerics.R
\name{separate_factors_numerics}
\alias{separate_factors_numerics}
\title{Split df into one of only factors and one of only numerics.}
\usage{
separate_factors_numerics(data, strings_to_factors = TRUE, verbose = FALSE)
}
\arguments{
\item{data}{A dataframe or matrix.}

\item{strings_to_factors}{Convert character strings to factors if True,
otherwise ignore character string variables.}

\item{verbose}{If TRUE output additional details during execution.}
}
\value{
Results list.

Convert strings to factors as well.
}
\description{
Split df into one of only factors and one of only numerics.
}
