% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_na.R
\name{sum_na}
\alias{sum_na}
\title{Get missingness for each column}
\usage{
sum_na(x)
}
\arguments{
\item{x}{Vector, matrix, or dataframe}
}
\description{
Function for getting total number missing values for vector
}

