% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cv_folds.R
\name{create_cv_folds}
\alias{create_cv_folds}
\title{Stratified CV to insure balance (by one grouping variable, Y)}
\usage{
create_cv_folds(V, Y, verbose = F)
}
\arguments{
\item{V}{number of folds}

\item{Y}{Outcome variable. If binary will be used for stratification.}
}
\value{
Vector of fold assignments.
}
\description{
Stratified CV to insure balance (by one grouping variable, Y)
}

