% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_tmle2.R
\name{estimate_tmle2}
\alias{estimate_tmle2}
\title{Get TMLE estimate: E[Y | A = 1, W].}
\usage{
estimate_tmle2(Y, A, W, family, delta = rep(1, length(Y)), Q.lib, g.lib,
  id = 1:length(Y), Qbounds = NULL, gbound = 0.025, alpha = 0.995,
  fluctuation = "logistic", verbose = F)
}
\arguments{
\item{Y}{Outcome variable}

\item{A}{Treatment indicator}

\item{W}{Dataframe of adjustment covariates}

\item{family}{Outcome family - binomial or gaussian}

\item{delta}{Indicator of missing outcome or treatment assignment. 1 - observed, 0 - missing.}

\item{Q.lib}{SuperLearner library for estimating Q (potential outcome)}

\item{g.lib}{SuperLearner library for estimating g (propensity score)}

\item{id}{Optional subject-level identifier.}

\item{Qbounds}{Bounds on Q}

\item{gbound}{Bounds on G}

\item{alpha}{TBD, from TMLE package}

\item{fluctuation}{Only logistic is currently supported.}

\item{verbose}{If true output extra information during execution.}
}
\description{
Get TMLE estimate: E[Y | A = 1, W].
}

