% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportLatex.R
\name{exportLatex}
\alias{exportLatex}
\title{Export varImpact results as Latex tables}
\usage{
exportLatex(impact_results, outname = "", dir = ".", digits = 4, ...)
}
\arguments{
\item{impact_results}{Result object from previous varImpact() call.}

\item{outname}{(Optional) String that is prepended to filenames.}

\item{dir}{(Optional) Directory to save the results, defaults to current directory.}

\item{digits}{Digits to round numbers, passed through to xtable.}

\item{...}{Additional parameters passed to print.xtable().}
}
\description{
Outputs results from varImpact() into three Latex tables: consistent results,
all results, and per-fold results.
}
\details{
Creates three Latex table files:
 \itemize{
   \item varimpConsistent.tex - the ``consistent'' significant results, meaning those with consistent categories chosen as comparison groups among factors and consistent ordering for numeric variables.
   \item varimpAll.tex - the file with cross-validated average variable impacts ordered by statistical significance.
   \item varimpByV.tex - the comparison levels used within each validation sample.  Either integer ordering of factors or short-hand for percentile cut-off (0-1 is the 10th percentile, 10+ is the 100th percentile)
}
}
\seealso{
\code{\link[varImpact]{varImpact}}
}

