% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.varImpact.R
\name{print.varimpact}
\alias{print.varimpact}
\title{Custom printing of the varimpact results.}
\usage{
\method{print}{varimpact}(x, ...)
}
\arguments{
\item{x}{Results object from varimpact.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Shows the significant and consistent results by default. If there are no
consistent results it shows all results.
}
