% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-by-level.R
\name{results_by_level}
\alias{results_by_level}
\title{Aggregate the results_by_fold_and_level df into a results_by_level df.}
\usage{
results_by_level(results_by_fold_and_level, verbose = FALSE)
}
\arguments{
\item{results_by_fold_and_level}{Dataframe containing the VIM results for
all levels of each variable across all CV folds.}

\item{verbose}{If true, display extra output.}
}
\description{
Aggregate the results_by_fold_and_level df into a results_by_level df.
}
