% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_estimate_q.R
\name{tmle_estimate_q}
\alias{tmle_estimate_q}
\title{purpose: estimate Q=E(Y |Z, A,W) data-adaptively,}
\usage{
tmle_estimate_q(
  Y,
  A,
  W,
  Delta,
  Q = NULL,
  Qbounds,
  Qform = NULL,
  maptoYstar,
  SL.library,
  family,
  id = 1:length(Y),
  V = 10,
  verbose = F
)
}
\arguments{
\item{Y}{outcome}

\item{A}{treatment indicator (1=treatment, 0=control)}

\item{W}{baseline covariates}

\item{Delta}{missingness indicator}

\item{Q}{optional externally estimated values for Q}

\item{Qbounds}{bounds for predicted values}

\item{Qform}{optional regression formula to use for glm if}

\item{maptoYstar}{if TRUE, using logistic fluctuation for bounded, continuous outcomes}

\item{SL.library}{library of prediction algorithms for Super Learner}

\item{family}{regression family}

\item{id}{subject identifier}

\item{V}{number of folds for SuperLearner}

\item{verbose}{Set T for extra output}
}
\description{
unless super learner not available, or user specifies
initial values or a regression formula
arguments:
}
